requirements.txt
amqp==2.5.0
anyjson==0.3.3
audioread==2.1.5
autopep8==1.4.4
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
billiard==3.6.0.0
bleach==1.4.3
boto==2.49.0
celery==4.3.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cloudinary==1.11.0
cssmin==0.2.0
decorator==4.2.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.11.11
django-advanced-filters==1.0.7.1
django-appconf==1.0.3
django-assets==0.12
django-auditlog==0.4.5
django-braces==1.4.0
django-bulk-update==2.2.0
django-classy-tags==0.8.0
django-compressor==2.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-easy-pdf==0.1.1
django-extensions==2.1.3
django-filer==1.4.4
django-filter==1.1.0
django-froala-editor==2.9.3
django-js-asset==1.2.1
django-jsonfield==1.2.0
django-mptt==0.9.1
django-nested-formset==0.1.4
django-polymorphic==2.0.3
django-postgres-extensions==0.9.3
django-qrcode==0.3
django-redis==4.10.0
django-render-partial==0.3
django-reversion==3.0.4
django-simple-history==2.7.3
django-taggit==0.23.0
django-transmeta==0.7.3
django-wkhtmltopdf==3.2.0
django-wysiwyg-redactor==0.5.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
easy-thumbnails==2.6
et-xmlfile==1.0.1
funcsigs==1.0.2
gprof2dot==2017.9.19
gunicorn==19.6.0
html5lib==1.0.1
httplib2==0.11.3
idna==2.8
jdcal==1.4.1
Jinja2==2.10.1
jsmin==2.2.2
kombu==4.6.3
loading==0.0.1
Markdown==2.6.6
MarkupSafe==1.1.1
mock==3.0.5
openpyxl==2.5.3
Pillow==3.3.0
pkg-resources==0.0.0
psycopg2==2.7.6
pycodestyle==2.5.0
pycparser==2.18
pydub==0.21.0
Pygments==2.4.2
PyJWT==1.7.1
PyPDF2==1.26.0
python-dateutil==2.6.0
python-memcached==1.59
pytz==2019.1
rcssmin==1.0.6
redis==3.2.1
reportlab==3.5.23
requests==2.22.0
rjsmin==1.1.0
simplejson==3.6.5
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
static3==0.7.0
typing==3.7.4
Unidecode==1.0.23
Unipath==1.0
urllib3==1.22
vine==1.3.0
webassets==0.12.1
webencodings==0.5.1
wget==3.2
whitenoise==4.1.2
xhtml2pdf==0.2.2
